arm: add initial Microchip SparX-5 support
Add initial support for the Microchip SparX-5 SoC.
It is a dedicated network switch SoC, but features 2 Cortex-A53 cores
running at 1GHz fixed as well as Cortex-M3 MCU.

Unlike most ARMv8 SoC-s, there is no BootROM, ATF nor any other FW
running on the SoC leaving U-boot in a unique position to provide
everything in a open-source form which is possible due to the datasheet
as well as registers being open to everybody.

Provide core register headers, DRAM training code and boot0 which is
required to put the Cortex-M3 sub-CPU into reset to use its SRAM for
initial setup until DRAM is trained.

Support for SMP by spin-tables is included as well, I could not get
PSCI CPU_ON to work so spin-tables only for now.

Like Ocelot and similar, we have to handle the CS-es manually as they
are not just plain GPIO-s.

Provide support for doing that as part of the architecture support as
this SoC can only boot from SPI-NOR so it will affect all boards.

This is heavily based on the vendor BSP U-Boot.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>