pinctrl: mscc: add SparX-5 GPIO and pinctrl
Add basic GPIO and pinctrl support for the 64 pins that Microchip
SparX-5 exposes from the SoC.

Add the required pinctrl symbols to be autoselected by ARCH_SPARX5 so the
driver can be autoselected once the ARCH_SPARX5 is selected.

Functions have been copied from Linux, so they are compatible with Linux
DTS-es.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>