net: mscc: add MIIM MDIO DM driver
MIIM is a common MDIO controller used on MSCC SoC-s, however it currently
does not have a DM driver but rather switch drivers pass platform data
and invoke the driver ops manually.

This leads to a lot of code duplication instead, so add a DM driver
for MIIM.

It is based on the Linux driver as well as the existing driver.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>